move app.js; rm default logo.svg; update manifest.json;
