+++ app/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "TMDB Search App",
+  "name": "TMDB Search React App Project",
   "icons": [
     {
       "src": "favicon.ico",
